Program Files
• Using stacks, write a mini infix expression parser.
• Your parser should parse an infix expression that supports the
following arithmetic and logical
operators with the specified precedencies:
! //logical not 8 ! 1 // = 0 (false)
^ //power 7 2^3 // 8
*, /, % //arithmetic 6 6 * 2 // 12
+, - //arithmetic 5 6 – 2 //4
>, >=, <, <= //comparison 4 6 > 5 // 1(true)
==, != //equality comparison 3 6!=5 // 1(true)
&& //logical and 2 6>5 && 4>5 //0(false)
|| //logical OR 1 1 || 0 //1 (true)
You may assume that all operands are integers. The result of a
comparison is a boolean (e.g.
6==6 is true). However, C++ compiler allows a boolean can be converted
to an integer according
to this logic: true =1 and false = 0. An integer can be converted to a
boolean according to this
logic: a number that is equal to 0 is false. Otherwise, it is true.
• You can be inspired by the postfix evaluator as well as the infix to
postfix convertor that we have
studied in class (the source code is on Blackboard). However, make sure
you come up with an
efficient algorithm.
• Your program should read several expressions the user has written in a
file, and evaluate them.
• Your program should be flexible with the expressions written by the
user in the file. For
instance, 1+2 is the same as 1 + 2. The user should not worry about
writing the spaces between
operands and operators.
• Your program should check for invalid expressions, and produce
meaningful error messages
when necessary. Here are a few examples: “invalid operator”, “there are
more operands than
operators”. Also tell the user where the error was detected (the
position of the character in the
expression).